Перенеси лишний код в общие модули
