+++ js/game.js
@@ -1,14 +1,19 @@
-// import {gameRender, statsRender} from './renderModule.js';
-// import renderHeader from './header.js';
-// import renderGreeting from './greeting.js';
+import renderHeader from './header.js';
+import renderGreeting from './greeting.js';
 import renderGame1 from './game-1.js';
 import renderGame2 from './game-2.js';
 import renderGame3 from './game-3.js';
-import {hasNextLevel} from './state.js';
-import renderStats from './stats.js';
+import {hasNextLevel, resetTimer} from './state.js';
+import renderTotalStats from './stats.js';
 
 const renderGame = (state) => {
   if (hasNextLevel(state.level, state.questions)) {
+    renderHeader(state);
+    resetTimer(state);
+    const backButton = document.querySelector(`.back`);
+    backButton.addEventListener(`click`, () => {
+      renderGreeting();
+    });
     switch (state.questions[state.level].type) {
       case `single`:
         renderGame2(state, renderGame);
@@ -20,7 +25,7 @@ const renderGame = (state) => {
         renderGame3(state, renderGame);
     }
   } else {
-    renderStats(state);
+    renderTotalStats(state);
   }
 
 };
